1. Book model updated to avoid nullable columns 2. /add, /delete, /update to add, delete or update books using JSON sent over POST requests

+++ book.py
@@ -26,27 +26,42 @@ def add():
 @book_app.route('/all')
 def record_show():
     test_records = Book.query.all()
-    # for record in test_records:
-    #     print(str(record.id)+" "+record.title+" "+record.author+" "+record.genre+" "+record.notes)
+    for record in test_records:
+        print(str(record.id)+" "+record.title+" "+record.author+" "+record.genre+" "+record.notes)
     return jsonify(response=str(len(test_records))), HTTPStatus.OK
 
 
 @book_app.route('/delete', methods=['POST'])
 def record_delete():
     content = request.get_json()
-    record = Book.query.filter_by(title=content['title']).first()
-    db.session.delete(record)
-    db.session.commit()
-    return jsonify(response='Successfully deleted'), HTTPStatus.OK
+    if 'title' in content:
+        record = Book.query.filter_by(title=content['title']).first()
+        db.session.delete(record)
+        db.session.commit()
+        return jsonify(response='Successfully deleted'), HTTPStatus.OK
+    else:
+        return jsonify(response='Needs a book title to delete'), HTTPStatus.BAD_REQUEST
 
 
 @book_app.route('/update', methods=['POST'])
 def record_update():
     content = request.get_json()
-    record = Book.query.filter_by(title=content['title']).first()
-    record.author = content['author']
-    record.published_on = content['published_on']
-    record.notes = content['notes']
+    if 'title' in content:
+        record = Book.query.filter_by(title=content['title']).first()
+    if 'genre' in content:
+        record.genre = content['genre']
+    if 'author' in content:
+        record.author = content['author']
+    if 'published_on' in content:
+        record.published_on = content['published_on']
+    if 'notes' in content:
+        record.notes = content['notes']
+    if 'lent_to' in content:
+        record.lent_to = content['lent_to']
+    if 'to_be_returned_on' in content:
+        record.to_be_returned_on = content['to_be_returned_on']
+    if 'is_returned' in content:
+        record.is_returned = content['is_returned']
     db.session.commit()
     return jsonify(response='Successfully updated'), HTTPStatus.OK
